refactor(consulcatalog): reorganize code.
